backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
networkx>=3.2
google-generativeai>=0.3.0
rapidfuzz>=3.5.0
pydantic>=2.5.0
python-dotenv>=1.0.0
groq>=0.9.0
